rename file for Draw component
